feat: LineDomainService 제거
